+++ rng.py
@@ -21,7 +21,7 @@ class Multithreaded_Standard_Normal:
         self.step = np.ceil(n / self.threads).astype(np.int_)
 
         def _fill(random_state, out, first, last):
-            random_state.standard_normal(out=out[first:last])
+            random_state.standard_normal(out=out[first:last], dtype=np.float32)
 
         futures = {}
         for i in range(self.threads):
@@ -72,7 +72,7 @@ def generate_perlin_noise_2d(
     grid = np.mgrid[0:res[0]:delta[0], 0:res[1]:delta[1]]\
              .transpose(1, 2, 0) % 1
     # Gradients
-    angles = 2 * np.pi * rng.random((res[0] + 1, res[1] + 1))
+    angles = 2 * np.float32(np.pi) * rng.random((res[0] + 1, res[1] + 1), dtype=np.float32)
     gradients = np.dstack((np.cos(angles), np.sin(angles)))
     if tileable[0]:
         gradients[-1, :] = gradients[0, :]
